+++ app/scripts/filter_directives.js
@@ -5,7 +5,7 @@
  */
 import _ from "lodash"
 import settings from "@/settings"
-import { regescape, translateAttribute } from "@/util"
+import { regescape, locAttribute } from "@/util"
 
 const korpApp = angular.module("korpApp")
 
@@ -116,8 +116,7 @@ korpApp.directive("globalFilter", [
 
             scope.isSelectedList = (value) => scope.selected.includes(value)
 
-            scope.translateAttribute = (value) =>
-                translateAttribute(scope.lang, scope.attrDef.settings.translation, value)
+            scope.translateAttribute = (value) => locAttribute(scope.attrDef.settings.translation, value, scope.lang)
         },
     }),
 ])